This improves performance of https://github.com/nazar-pc/abundance/pull/103 further:
```
flipper/direct          time:   [77.357 ns 77.609 ns 77.909 ns]
                        thrpt:  [12.835 Melem/s 12.885 Melem/s 12.927 Melem/s]
flipper/transaction     time:   [92.766 ns 93.097 ns 93.498 ns]
                        thrpt:  [10.695 Melem/s 10.741 Melem/s 10.780 Melem/s]
```

There are also a few small fixes and improves docs.